+++ tests/test.bash
@@ -4,8 +4,9 @@
 
 SCRIPT=$0;
 SCRIPT_NAME=`basename $SCRIPT`;
+PROJECT_DIR=$(readlink -f "`dirname $SCRIPT`/..");
+TEST_TMP_DIR="$PROJECT_DIR/test-temp";
 WORK_DIR=`pwd`;
-TEST_TMP_DIR="$WORK_DIR/test-temp";
 
 ERROR=;
 
@@ -43,48 +44,22 @@ mkdir -p $TEST_TMP_DIR;
 cd $TEST_TMP_DIR;
 
 # Prepare for test (if running on CI)
-if [ ! -z $CI_RUN ]; then
-    echo "Running as in CI environment";
-    export ALWAYS_ANSWER_YES=1;
+source "$PROJECT_DIR/tests/ci.bash";
 
-    if ! command -v "odoo-install" >/dev/null 2>&1 || ! command -v "odoo-helper" >/dev/null 2>&1; then
-        echo "Seems that odoo-helper-scripts were not installed correctly!";
-        echo "PATH: $PATH";
-        echo "Current path: $(pwd)";
-        echo "Home var: $HOME";
-        echo "";
-        if [ -f $HOME/odoo-helper.conf ]; then
-            echo "User conf: ";
-            echo "$(cat $HOME/odoo-helper.conf)";
-        else
-            echo "User conf not found!";
-        fi
-        echo "";
-        echo "Content of ~/.profile file:";
-        echo "$(cat $HOME/.profile)";
-        echo "";
-        echo "Content of ~/.bashrc file:";
-        echo "$(cat $HOME/.bashrc)";
-        echo "";
-        echo "Content of ~/.bash_profile file:";
-        echo "$(cat $HOME/.bash_profile)";
-        echo "";
-        
-    fi
-    sudo pip install --upgrade pytz
-fi
+# import odoo-helper common lib to allow colors in test output
+source $(odoo-helper system lib-path common);
+allow_colors;
 
 #
 # Start test
 # ==========
 #
-
-echo "
+echo -e "${YELLOWC}
 =================================================
 Test install of odoo version 7.0
 Also install dependencies and configure postgresql
 ==================================================
-"
+${NC}"
 odoo-install --install-dir odoo-7.0 --branch 7.0 --extra-utils --install-and-conf-postgres --install-sys-deps \
     --conf-opt-xmlrpc_port 8369 --conf-opt-xmlrpcs_port 8371
 cd odoo-7.0
@@ -93,12 +68,12 @@ cd odoo-7.0
 # Note, thant Odoo this odoo install uses virtual env (venv dir)
 # Also You will find there odoo-helper.conf config file
 
-echo "
+echo -e "${YELLOWC}
 =================================
 Test database management features
 (create, list, and drop database)
 =================================
-"
+${NC}"
 
 # create test database if it does not exists yet
 if ! odoo-helper db exists my-test-odoo-database; then
@@ -108,17 +83,28 @@ fi
 # list all odoo databases available for this odoo instance
 odoo-helper db list
 
+# backup database
+backup_file=$(odoo-helper db backup my-test-odoo-database);
+
 # drop test database if it exists
 if odoo-helper db exists my-test-odoo-database; then
     odoo-helper db drop my-test-odoo-database;
 fi
 
+# restore dropped database
+odoo-helper db restore my-test-odoo-database $backup_file;
 
-echo "
+# ensure that database exists
+odoo-helper db exists my-test-odoo-database
+
+# drop database egain
+odoo-helper db drop my-test-odoo-database;
+
+echo -e "${YELLOWC}
 =================================
 Test 'odoo-helper server' command
 =================================
-"
+${NC}"
 # So now You may run local odoo server (i.e openerp-server script).
 # Note that this command run's server in foreground.
 odoo-helper server --stop-after-init  # This will automaticaly use config file: conf/odoo.conf
@@ -142,11 +128,11 @@ odoo-helper server stop
 odoo-helper server status
 
 
-echo "
+echo -e "${YELLOWC}
 ============================================================
 Fetch and test 'https://github.com/katyukha/base_tags' addon
 ============================================================
-"
+${NC}"
 # Let's install base_tags addon into this odoo installation
 odoo-helper fetch --github katyukha/base_tags --branch master
 
@@ -169,12 +155,12 @@ odoo-helper --use-unbuffer test --create-test-db -m base_tags -m product_tags
 # go back to directory containing our projects (that one, where odoo-7.0 project is placed)
 cd ../../
 
-echo "
+echo -e "${YELLOWC}
 ========================================================================
 Install odoo 8.0, fetch and run tests for OCA addon 'project_sla'
-Alson install python package 'suds' in virtual env of this odoo instance
+Also install python package 'suds' in virtual env of this odoo instance
 ========================================================================
-"
+${NC}"
 # Let's install odoo of version 8.0 too here.
 odoo-install --install-dir odoo-8.0 --branch 8.0 --extra-utils\
     --conf-opt-xmlrpc_port 8369 --conf-opt-xmlrpcs_port 8371 --conf-opt-longpolling_port 8372
@@ -196,13 +182,13 @@ odoo-helper fetch -p suds  # this installs 'suds' python package
 # Show addons status for this project
 odoo-helper --use-unbuffer addons status
 
-echo "
+echo -e "${YELLOWC}
 ===============================================================================
 Go back to Odoo 7.0 instance, we installed at start of test
 and fetch and install there aeroo reports addon with it's dependency 'aeroolib'
 After this, generate requirements list.
 ===============================================================================
-"
+${NC}"
 
 # and as one more example, let's install aeroo-reports with dependancy to aeroolib in odoo 7.0
 cd ../odoo-7.0
@@ -210,11 +196,11 @@ odoo-helper fetch --github gisce/aeroo -n aeroo
 odoo-helper fetch -p git+https://github.com/jamotion/aeroolib#egg=aeroolib
 odoo-helper generate_requirements
 
-echo "
+echo -e "${YELLOWC}
 ==========================
 Install and check Odoo 9.0 
 ==========================
-"
+${NC}"
 
 # got back to test root and install odoo version 9.0
 cd ../;